MfgTool.exe - v1.6.2.025 : Prepare for source code release.
- Removed code that was no longer being used.
- Moved code obtained from public domain into Libs/Public
- Removed Libs/WinSupport from MfgTool solution
- Removed Libs/OtpAccessPitc from MfgTool solution
- Removed Freescale copyright from public domain files.

+++ Apps/MfgTool.exe/OpMxRomUpdate.cpp
@@ -56,7 +56,7 @@ COpMxRomUpdate::COpMxRomUpdate(CPortMgrDlg *pPortMgrDlg, usb::Port *pUSBPort, CO
 	}
 
 	CStringT<char,StrTraitMFC<char> > uclString;
-	commandFile.Read(uclString.GetBufferSetLength(commandFile.GetLength()), commandFile.GetLength());
+	commandFile.Read(uclString.GetBufferSetLength((int)commandFile.GetLength()), (unsigned int)commandFile.GetLength());
 	uclString.ReleaseBuffer();
 
 	if ( m_UclNode.Load(A2T(uclString)) != NULL )
